Simplify header.
