+++ src/service.ts
@@ -243,7 +243,7 @@ export class DebuggerService implements IDebugger, IDisposable {
     }
 
     if (this.isStarted || autoStart) {
-      this._changeTitle();
+      this._model.title = this.isStarted ? this.session?.connection?.name : '-';
     }
 
     this._model.breakpoints.restoreBreakpoints(bpMap);
@@ -381,14 +381,6 @@ export class DebuggerService implements IDebugger, IDisposable {
   async dumpCell(code: string) {
     const reply = await this.session.sendRequest('dumpCell', { code });
     return reply.body;
-  }
-
-  /**
-   * Emit session which have set enabled debugging
-   */
-  private _changeTitle() {
-    const title = this.isStarted ? this.session?.connection?.name : '-';
-    this._model.title = title;
   }
 
   /**
